# backend/swagger.yaml
swagger: "2.0"
info:
  description: "A web-based  judging  application  that  allows  users,  namely Admin, to host events and Participants, to create their project ideas or join an existing project and Judges, to be seamlessly involved in rating the assigned projects. Participants create and join the projects in an event. Eventually, judges are allowed to rate and admins can announce the winners of the projects in each event that are assigned to them. Each user has their roles and access to perform their respective operations described in the upcoming sections."
  version: "1.0.0"
  title: "SeamlessJudging"
host: "stark.cse.buffalo:3000"
schemes:
- "http"
paths:
...
  /api/v1/content/bulk-upload:
    post:
      tags:
      - "Content"
      summary: "Add entities in bulk"
      description: "Only admin can add in bulk"
      operationId: "bulkUpload"
      produces:
      - "application/json"
      parameters:
      - name: "Access-Token"
        in: "header"
        description: ""
        required: true
        type: "string"
      - in: "body"
        name: "body"
        description: "Bulk Upload entities with the help of a csv"
        required: true
        schema:
          type: "object"
          properties:
            upload_type:
              type: "string"
              example: "PROJECT"
            file:
              type: "string"
              example: "C://User/Desktop/projects.xlsx"
      responses:
          "201":
            description: "successfully added"
            schema:
              type: "object"
              properties:
                response_str:
                    type: "string"
                    example: "Bulk-Upload successful"
                response_data:
                  type: "object"
                  properties:
                    name:
                        type: "string"
                        example: "MS Project Development - 2022"
                    project_type_name:
                      type: "string"
                      example: "Development"
                    course_code:
                      type: "string"
                      example: "CSE-611"
                    team_size:
                      type: "integer"
                      example: 5
                    client_name:
                      type: "string"
                      example: "Alan Hunt"
                    client_company:
                      type: "string"
                      example: "UB"
                    client_email:
                      type: "string"
                      example: "ahunt@buffalo.edu"
                    project_link:
                      type: "string"
                      example: "http://stark.cse.buffalo.edu:8000"
                    status:
                      type: "string"
                      example: "FAILED"
                    result:
                      type: "string"
                      example: "Invalid course-code"
                    team:
                      type: "string"
                      example: "joh@abc.com,doe@abc.com"
                    
          "404":
            description: "Not Found"

    get:
      tags:
      - "Content"
      summary: "Retrieve Bulk-Upload templates"
      description: "Only admin retreive bulk-upload templates"
      operationId: "getBulkUploadTemplates"
      produces:
      - "application/json"
      parameters:
      - name: "Access-Token"
        in: "header"
        description: ""
        required: true
        type: "string"
      - name: "upload_type"
        in: "query"
        description: ""
        required: true
        type: "string"
      responses:
          "200":
            description: "successful"
            schema:
              type: "object"
              properties:
                response_str:
                    type: "string"
                    example: "Project Types details retrieved successfully"
                response_data:
                  type: "object"
                  properties:
                    file:
                      type: "string"
                      example: "template.xlsx"
          "404":
            description: "Not Found"
...